adding concept and classic

+++ llob/Types.h
@@ -3,7 +3,7 @@
 namespace llob {
 
 using Price = double;
-using InstrumentId = uint32_t;
+using InstrumentId = uint16_t;
 using Quantity = uint64_t;
 using OrderId = uint64_t;
